lgd/requirements.txt
beautifulsoup4==4.9.3
beniget==0.4.1
bs4==0.0.1
cachetools==4.2.4
calmjs.parse==1.3.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorlog==6.6.0
cycler==0.10.0
Cython==0.29.24
defusedxml==0.7.1
gast==0.5.2
google-api-core==2.3.0
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
humanize==3.13.1
idna==3.2
imageio==2.9.0
kiwisolver==1.1.0
matplotlib==3.4.2
networkx==2.6.2
numpy==1.21.1
odfpy==1.4.1
Pillow==8.3.1
ply==3.11
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.7.1
pyparsing==2.4.7
pytesseract==0.3.8
python-dateutil==2.8.2
pythran==0.9.12.post1
PyWavelets==1.1.1
requests==2.26.0
rsa==4.8
scikit-image==0.18.2
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
tifffile==2021.7.30
tqdm==4.62.3
urllib3==1.26.6
xlsx2csv==0.7.8
